实现了用线程管理的链表化，包括current之内的很多接口都可以使用了，此版本编译运行都ok

+++ include/stddef.h
@@ -4,10 +4,12 @@
 #undef NULL
 #define NULL ((void *)0)
 
+#include "kernel.h"
+
 enum {
 	false	= 0,
 	true	= 1
 };
 
-#define offsetof(TYPE, MEMBER) ((size_t) &((TYPE *)0)->MEMBER)
+#define offsetof(TYPE, MEMBER) ((unsigned long) &((TYPE *)0)->MEMBER)
 #endif
